Add layout stories. #1279
