+++ services/ui/src/layouts/StatusLayout/index.js
@@ -2,6 +2,9 @@ import GlobalStlyes from 'layouts/GlobalStyles';
 import Header from 'components/Header';
 import { bp } from 'lib/variables';
 
+/**
+ * The status layout includes the Lagoon UI header and grey box to wrap content.
+ */
 const StatusLayout = ({ children }) => (
   <GlobalStlyes>
     <Header />
